リファクタリング: Selectorをmoduleに。メソッド名をオプションと合わせる
