+++ 04.class/memo.js
@@ -1,10 +1,10 @@
 import fs from "fs";
 import minimist from "minimist";
 import sqlite3 from "sqlite3";
+import { choice } from "./module/selector_module.js";
 import { close } from "./module/sqlite3_module.js";
-import { index, show } from "./module/view_module.js";
+import { list, reference } from "./module/view_module.js";
 import { Memo } from "./class/memo_class.js";
-import { Selector } from "./class/selector_class.js";
 
 const DB = "memos.db";
 
@@ -15,13 +15,16 @@ async function main() {
   try {
     if (options.l) {
       const memos = await Memo.all(db);
-      index(memos);
+      list(memos);
     } else if (options.r) {
-      const choiced_memo = await choice(db, "Choose a note you want to show:");
-      show(choiced_memo);
+      const choiced_memo = await choice(
+        await Memo.all(db),
+        "Choose a note you want to reference:"
+      );
+      reference(choiced_memo);
     } else if (options.d) {
       const choiced_memo = await choice(
-        db,
+        await Memo.all(db),
         "Choose a note you want to delete:"
       );
       await Memo.destroy(db, choiced_memo);
@@ -39,14 +42,6 @@ async function main() {
   } finally {
     await close(db);
   }
-}
-
-// DRY method
-async function choice(db, msg = "") {
-  const memos = await Memo.all(db);
-  const selector = new Selector(memos);
-  const choiced_memo = await selector.choice(msg);
-  return choiced_memo;
 }
 
 main();
